TEST: 공통변수 분리 & 회전기능 테스트코드 추가
